// src/app/core/models/province.model.ts
/** @deprecated */
export interface ProvinceModel {
    id:number;
    filter_by_answer_name: string;
    filter_by_answer_value: string;
    value: string;
    label: string;
    irrigated_dry: number;
    irrigated_wet: number;
    rainfed_dry: number;
    rainfed_wet: number;
    corn_avg: number;
    irrigated_1stq: number;
    irrigated_2ndq: number;
    irrigated_3rdq: number;
    irrigated_4thq: number;
    rainfed_1stq: number;
    rainfed_2ndq: number;
    rainfed_3rdq: number;
    rainfed_4thq: number;
}